+++ src/app/map/map.component.ts
@@ -11,6 +11,7 @@ import * as messageType from '../shared/messageType';
 import * as esri from 'esri-leaflet';
 import { Workflow } from '../shared/interfaces/workflow/workflow';
 import { LoaderService } from '../shared/services/loader.service';
+import { AppService } from '../shared/services/app.service';
 
 @Component({
   selector: 'app-map',
@@ -24,6 +25,7 @@ export class MapComponent implements OnInit {
   public currentZoom: number = 4;
   public latestDischarge: any;
 	public workflowLayers = [] as any;
+	public activeWorkflowLayers:any;
   public selectedFeature: any;
   public marker: L.Marker;
   public basin: any;
@@ -35,10 +37,11 @@ export class MapComponent implements OnInit {
   public selectedPopup: any;
   public selectedSite: any
   public streamgageLayer: any;
+  public streamgageLayerStatus: boolean;
   public workflowData: any;
   public count: number = 0;
   constructor(public _mapService: MapService, private _configService: ConfigService, private _http:
-     HttpClient, private _workflowService: WorkflowService, public toastr: ToastrService, private _loaderService: LoaderService) { 
+     HttpClient, private _workflowService: WorkflowService, public toastr: ToastrService, private _loaderService: LoaderService, private _appService: AppService) { 
     this.configSettings = this._configService.getConfiguration();
     this.messager = toastr;
   }
@@ -56,6 +59,20 @@ export class MapComponent implements OnInit {
 
     // Add basemap
     this._mapService.SetBaselayer(this._mapService.chosenBaseLayerName);
+
+    // Get available workflow layers / Get active workflow layers
+    this.workflowLayers = this._mapService.workflowLayers;
+    this._mapService.activeWorkflowLayers.subscribe(activeLayers => {
+      this.activeWorkflowLayers = activeLayers;
+    })
+
+    // Get streamgages toggle status
+    this._mapService.streamgageLayerStatus.subscribe((status: boolean) => {
+      this.streamgageLayerStatus = status;
+      if (this.streamgageLayerStatus) {
+        this.setBbox();
+      }
+    })
 
     // Add scale bar
     this._mapService.scale.addTo(this._mapService.map);
@@ -110,8 +127,8 @@ export class MapComponent implements OnInit {
     // On map zoom, set current zoom, display gages
     this._mapService.map.on('zoomend',(evt) => {
       this.currentZoom = evt.target._zoom;
+      this._mapService.setCurrentZoomLevel(evt.target._zoom);
       this.setBbox();
-      this.checkAvailableLayers();
     }) 
     // On map drag, display gages
     this._mapService.map.on('dragend',() => {
@@ -120,23 +137,26 @@ export class MapComponent implements OnInit {
     // Subscribe to workflow
     this._workflowService.selectedWorkflow.subscribe((res) => {
       this.selectedWorkflow = res;
-      this.checkAvailableLayers();
+      if (this.selectedWorkflow) {
+        if(this.selectedWorkflow.title != "Delineation") {
+          this.checkAvailableLayers();
+        }
+      }
       if (!this.selectedWorkflow){
-        Object.keys(this.workflowLayers).forEach(layerName => {
-          this.removeLayer(this.workflowLayers[layerName]); // No workflow is selected; remove all workflow overlayers
-        });
-
+        this.removeWorkFlowLayers();
       }
     });
 
     //Subscribe to the form data
     this._workflowService.formData.subscribe(data => {
       this.workflowData = data;
-      this.checkAvailableLayers();
+      if (this.workflowData) {
+        if (this.workflowData.title == "Delineation") {
+          this.checkAvailableLayers();
+        }
+      }
       if (!this.workflowData) {
-        Object.keys(this.workflowLayers).forEach(layerName => {
-          this.removeLayer(this.workflowLayers[layerName]); // No workflow is selected; remove all workflow overlayers
-        });
+        this.removeWorkFlowLayers();
       } 
     });
 
@@ -148,20 +168,30 @@ export class MapComponent implements OnInit {
     this.loadLayers();
   }
 
+  public removeWorkFlowLayers(){
+    Object.keys(this.workflowLayers).forEach(layerName => {
+      this._appService.setLayerVisibility(layerName);
+      this._mapService.removeWorkflowLayers(layerName);
+      this.removeLayer(this.workflowLayers[layerName]); // No workflow is selected; remove all workflow overlayers
+    });
+  }
+
   public checkAvailableLayers(){
     if (this.selectedWorkflow) {
       switch (this.selectedWorkflow.title) {
         case "Delineation":
-          if (this.workflowData && this.workflowData.steps) {
-            this.workflowData.steps[0].options.forEach(o => {
-              if (o.text == "NLDI Delineation" && o.selected == true) {
-                this.addLayers('NHD');
-              }
-            });
+          if (!this.activeWorkflowLayers.length) {
+            if (this.workflowData && this.workflowData.steps || this.activeWorkflowLayers.name != "NHD Flowlines") {
+              this.workflowData.steps[0].options.forEach(o => {
+                if (o.text == "NLDI Delineation" && o.selected == true) {
+                  this.addLayers('NHD Flowlines');
+                }
+              });
+            }
           }
           break;
         case "Fire Hydrology - Query Basin":
-          this.addLayers('NHD');
+          this.addLayers('NHD Flowlines');
           this.addLayers('Archived WildFire Perimeters');
           this.addLayers('Active WildFire Perimeters');
           this.addLayers('MTBS Fire Boundaries');
@@ -204,16 +234,31 @@ export class MapComponent implements OnInit {
   }
 
   public addLayers(layerName: string) {
+    this.configSettings.workflowLayers.forEach((layer: any) => {
+      if (layer.name === layerName) {
+        layer.visible = true;
+        this._mapService.setWorkflowLayers(layer);
+      }
+    }); 
     this.workflowLayers[layerName].addTo(this._mapService.map);
   }
 
   public setBbox(){
-    if (this.streamgageLayer !== undefined) this._mapService.map.removeLayer(this.streamgageLayer);
+    if (this.streamgageLayer !== undefined) {
+      this._mapService.map.removeLayer(this.streamgageLayer);
+      this.configSettings.overlays.forEach((overlay: any) => {
+        if (overlay.name === "Streamgages") {
+            overlay.visible = false;
+        }
+      });
+    }
     if (this.currentZoom >= 8) {
       var bBox = this._mapService.map.getBounds();
       var ne = bBox.getNorthEast(); // LatLng of the north-east corner
       var sw = bBox.getSouthWest(); // LatLng of the south-west corder
-      this._mapService.setStreamgages(sw.lng, ne.lng, sw.lat, ne.lat)
+      if (this.streamgageLayerStatus) {
+        this._mapService.setStreamgages(sw.lng, ne.lng, sw.lat, ne.lat)
+      }
     }
   }
   
@@ -244,6 +289,13 @@ export class MapComponent implements OnInit {
           return L.marker(latlng,{icon: MyIcon});
         }
       }).addTo(this._mapService.map);
+      this.configSettings.overlays.forEach((overlay: any) => {
+        if (overlay.name === "Streamgages") {
+          overlay.visible = true;
+        }
+      });
+      // setting streamgages layer in map services
+      this._mapService.setStreamgagesLayer(this.streamgageLayer)
     }
   }